% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{adjust.cohorts}
\alias{adjust.cohorts}
\title{Delivers data for Adjust Cohorts. Refer to the KPI service docs under https://docs.adjust.com/en/kpi-service/
together with this help entry.}
\usage{
adjust.cohorts(app.token = NULL, tracker.token = NULL, ...)
}
\arguments{
\item{app.token}{pass it here or set it up once with \code{\link{set.app.token}}}

\item{tracker.token}{If you want data for a given parent tracker, pass its token.}

\item{start_date}{YYYY-MM-DD The start date of the selected period.}

\item{end_date}{YYYY-MM-DD The end date of the selected period.}

\item{kpis}{A vector of App KPIs. See KPI service documentation https://docs.adjust.com/en/kpi-service/ for a
list of supported App KPIs.}

\item{period}{The cohort period - one of day/week/month.}

\item{events}{A vector of event tokens for Event-specific queries.}

\item{sandbox}{Boolean request only sandbox data}

\item{countries}{A vector of ISO 3166 alpha-2 country names.}

\item{os_names}{A vector of OS names. See KPI service documentation https://docs.adjust.com/en/kpi-service/ for a
list of supported OS names.}

\item{device_types}{A vector of supported device types.}

\item{grouping}{A vector of supported grouping. E.g. \code{c('trackers', 'countries')}. For more on grouping, see the
KPI service docs.}
}
\description{
Delivers data for Adjust Cohorts. Refer to the KPI service docs under https://docs.adjust.com/en/kpi-service/
together with this help entry.
}

