% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{adjust.deliverables}
\alias{adjust.deliverables}
\title{Delivers data for Adjust App KPIs. Refer to the KPI service docs under https://docs.adjust.com/en/kpi-service/
together with this help entry.}
\usage{
adjust.deliverables(app.tokens = NULL, ...)
}
\arguments{
\item{app.tokens}{pass it here or set it up once with \code{\link{set.app.tokens}}}

\item{start_date}{YYYY-MM-DD The start date of the selected period.}

\item{end_date}{YYYY-MM-DD The end date of the selected period.}

\item{kpis}{A vector of App KPIs. See KPI service docs for more.}

\item{sandbox}{Boolean request only sandbox data}

\item{countries}{A vector of ISO 3166 alpha-2 country names.}

\item{os_names}{A vector of OS names. See KPI service docs for more.}

\item{device_types}{A vector of supported device types.}

\item{grouping}{A vector of supported grouping. E.g. \code{c('countries')}. For more on grouping, see the
KPI service docs.}
}
\description{
Delivers data for Adjust App KPIs. Refer to the KPI service docs under https://docs.adjust.com/en/kpi-service/
together with this help entry.
}
\examples{
adjust.deliverables() # perhaps the simplest query, it uses the default request parameters.
adjust.deliverables(countries=c('us', 'de')) # scope by countries.
}
