% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{adjust.enable.verbose}
\alias{adjust.enable.verbose}
\title{Enable the verbose setting. Doing this will print out additional meta data on the API requests.}
\usage{
adjust.enable.verbose()
}
\description{
Enable the verbose setting. Doing this will print out additional meta data on the API requests.
}
\seealso{
\code{\link{adjust.disable.verbose}}
}
