% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{adjust.events}
\alias{adjust.events}
\title{Delivers data for Adjust Event KPIs. Refer to the KPI service docs under https://docs.adjust.com/en/kpi-service/
together with this help entry.}
\usage{
adjust.events(app.tokens = NULL, ...)
}
\arguments{
\item{start_date}{YYYY-MM-DD The start date of the selected period.}

\item{end_date}{YYYY-MM-DD The end date of the selected period.}

\item{kpis}{A vector of App KPIs. See KPI service documentation https://docs.adjust.com/en/kpi-service/ for a
list of supported App KPIs.}

\item{events}{A vector of event tokens for Event-specific queries.}

\item{sandbox}{Boolean request only sandbox data}

\item{countries}{A vector of ISO 3166 alpha-2 country names.}

\item{os_names}{A vector of OS names. See KPI service documentation https://docs.adjust.com/en/kpi-service/ for a
list of supported OS names.}

\item{device_types}{A vector of supported device types.}

\item{grouping}{A vector of supported grouping. E.g. \code{c('trackers', 'countries')}. For more on grouping, see the
KPI service docs.}
}
\description{
Delivers data for Adjust Event KPIs. Refer to the KPI service docs under https://docs.adjust.com/en/kpi-service/
together with this help entry.
}
\examples{
adjust.events() # perhaps the simplest query, it uses the default request parameters.
adjust.events(countries=c('us', 'de')) # scope by countries.
}
