% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{adjust.setup}
\alias{adjust.setup}
\title{Convenience function for initiating a session with a user token and app tokens, generally required for the start of
an Adjust API session. These settings could also be overwritten by the `adjust.cohorts`, `adjust.deliverables`, etc.
function arguments.}
\usage{
adjust.setup(user.token = NULL, app.tokens = NULL)
}
\description{
Convenience function for initiating a session with a user token and app tokens, generally required for the start of
an Adjust API session. These settings could also be overwritten by the `adjust.cohorts`, `adjust.deliverables`, etc.
function arguments.
}
\seealso{
\code{\link{set.user.token}}, \code{\link{user.token}}, \code{\link{set.app.tokens}}, \code{\link{app.tokens}}
}
