% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{app.token}
\alias{app.token}
\title{Get the currently set app.token.}
\usage{
app.token()
}
\description{
Get the currently set app.token.
}
\seealso{
\code{\link{set.user.token}}, \code{\link{user.token}}, \code{\link{set.app.token}}, \code{\link{adjust.setup}}
}

