% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_asset_alerts}
\alias{catalog_asset_alerts}
\title{Available Asset Alerts}
\usage{
catalog_asset_alerts(assets = NULL, alerts = NULL)
}
\arguments{
\item{assets}{Vector of assets. By default all assets are returned.}

\item{alerts}{Character vector of asset alert names. By default all asset alerts are returned.}
}
\value{
Tibble of available asset alerts along with their descriptions, thresholds, and constituents.
}
\description{
Available Asset Alerts
}
