% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_asset_metrics}
\alias{catalog_asset_metrics}
\title{Available Asset Metrics}
\usage{
catalog_asset_metrics(metrics = NULL, reviewable = NULL)
}
\arguments{
\item{metrics}{Vector of metrics. By default, all metrics are returned.}

\item{reviewable}{Boolean to limit to human-reviewable metrics. By default, all metrics are returned.}
}
\value{
Tibble of available asset metrics along with information like description, category, and assets for which a metric is available.
}
\description{
Available Asset Metrics
}
