% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_assets}
\alias{catalog_assets}
\title{Available Assets}
\usage{
catalog_assets(assets = NULL)
}
\arguments{
\item{assets}{Vector of assets. By default all assets are returned.}
}
\value{
List of available assets along with information for them (metrics, markets, and exchanges) and time ranges of available data.
}
\description{
Available Assets
}
