% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_exchange_metrics}
\alias{catalog_exchange_metrics}
\title{Available Exchange Metrics}
\usage{
catalog_exchange_metrics(metrics = NULL, reviewable = NULL, pretty = FALSE)
}
\arguments{
\item{metrics}{Vector of metrics. By default all metrics are returned.}

\item{reviewable}{Limit to human-reviewable metrics. By default all metrics are returned.}

\item{pretty}{Human-readable formatting of JSON responses}
}
\value{
Tibble of available exchange metrics along with information for them like description, category, and exchanges for which a metric is available.
}
\description{
Available Exchange Metrics
}
