% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_asset_alerts}
\alias{catalog_full_asset_alerts}
\title{Supported Asset Alerts}
\usage{
catalog_full_asset_alerts(assets = NULL, alerts = NULL)
}
\arguments{
\item{assets}{Vector of assets. By default, all assets are returned.}

\item{alerts}{Vector of alert names. By default, all asset alerts are returned.}
}
\value{
Tibble of all supported asset alerts along with their descriptions, thresholds, and constituents.
}
\description{
Supported Asset Alerts
}
