% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_asset_metrics}
\alias{catalog_full_asset_metrics}
\title{Supported Asset Metrics}
\usage{
catalog_full_asset_metrics(metrics = NULL, reviewable = NULL)
}
\arguments{
\item{metrics}{Vector of metrics. By default, all metrics are returned.}

\item{reviewable}{Boolean to limit to human-reviewable metrics. By default, all metrics are returned.}
}
\value{
Tibble of supported asset metrics along with information for them
like description, category, and assets for which a metric is available.
}
\description{
Supported Asset Metrics
}
