% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_asset_pairs}
\alias{catalog_full_asset_pairs}
\title{Supported Asset Pairs}
\usage{
catalog_full_asset_pairs(pairs = NULL, pretty = FALSE)
}
\arguments{
\item{pairs}{Vector of asset pairs. By default, all asset pairs are returned.}

\item{pretty}{Human-readable formatting of JSON responses.}
}
\value{
Tibble of supported asset pairs along with information for them like
metrics and time ranges of available data.
}
\description{
Supported Asset Pairs
}
