% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_assets}
\alias{catalog_full_assets}
\title{Supported Assets}
\usage{
catalog_full_assets(assets = NULL)
}
\arguments{
\item{assets}{Vector of assets. By default, all assets are returned.}
}
\value{
Tibble of supported assets along with information for them like metrics, markets, exchanges, and time ranges of available data.
Metrics, markets, and exchanges outputted as list-columns.
}
\description{
Supported Assets
}
