% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_exchange_assets}
\alias{catalog_full_exchange_assets}
\title{Supported Exchange-Asset Pairs}
\usage{
catalog_full_exchange_assets(exchange_assets = NULL, pretty = FALSE)
}
\arguments{
\item{exchange_assets}{Vector of exchange-assets. By default, all exchange-asset pairs are returned.}

\item{pretty}{Human-readable formatting of JSON responses.}
}
\value{
Tibble of all supported exchange-asset pairs along with information for them
like metrics and time ranges of available data.
}
\description{
Supported Exchange-Asset Pairs
}
