% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_exchange_metrics}
\alias{catalog_full_exchange_metrics}
\title{Supported Exchange Metrics}
\usage{
catalog_full_exchange_metrics(metrics = NULL)
}
\arguments{
\item{metrics}{.
like description, category, and exchanges for which a metric is available.}
}
\description{
Supported Exchange Metrics
}
