% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_index_candles}
\alias{catalog_full_index_candles}
\title{Supported Index Candles}
\usage{
catalog_full_index_candles(indexes = NULL)
}
\arguments{
\item{indexes}{Vector of index names. By default, all indexes are returned.}
}
\value{
Tibble of all supported index candles along with the time ranges of available data per index candle duration.
}
\description{
Supported Index Candles
}
