% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_indexes}
\alias{catalog_full_indexes}
\title{Supported Indexes}
\usage{
catalog_full_indexes(indexes = NULL)
}
\arguments{
\item{indexes}{Vector of index names. By default, all indexes are returned.}
}
\value{
Tibble of all supported indexes along with time ranges of available data.
}
\description{
Supported Indexes
}
