% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_institution_metrics}
\alias{catalog_full_institution_metrics}
\title{Supported Institution Metrics}
\usage{
catalog_full_institution_metrics(metrics = NULL)
}
\arguments{
\item{metrics}{.}
}
\description{
Supported Institution Metrics
}
