% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full_v2.R
\name{catalog_full_v2_asset_metrics}
\alias{catalog_full_v2_asset_metrics}
\title{Supported Asset Metrics}
\usage{
catalog_full_v2_asset_metrics(
  assets = NULL,
  metrics = NULL,
  reviewable = NULL,
  page_size = 2000,
  paging_from = "start"
)
}
\arguments{
\item{assets}{Character vector of assets. By default, all assets are returned.}

\item{metrics}{Vector of metrics. By default, all metrics are returned.}

\item{reviewable}{Boolean to limit to human-reviewable metrics. By default, all metrics are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of all supported asset metrics, along with the time ranges of available data.
}
\description{
Supported Asset Metrics
}
