% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full_v2.R
\name{catalog_full_v2_index_levels}
\alias{catalog_full_v2_index_levels}
\title{Supported Index Levels}
\usage{
catalog_full_v2_index_levels(
  indexes = NULL,
  page_size = 2000,
  paging_from = "start"
)
}
\arguments{
\item{indexes}{Character vector of indexes. By default, all indexes are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of all supported index levels, along with time ranges of available data.
}
\description{
Supported Index Levels
}
