% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full_v2.R
\name{catalog_full_v2_market_implied_volatility}
\alias{catalog_full_v2_market_implied_volatility}
\title{Supported Market Implied Volatility}
\usage{
catalog_full_v2_market_implied_volatility(
  markets = NULL,
  exchange = NULL,
  type = NULL,
  base = NULL,
  quote = NULL,
  asset = NULL,
  symbol = NULL,
  page_size = 2000,
  paging_from = "start"
)
}
\arguments{
\item{markets}{Character vector of markets. By default, all markets are returned.}

\item{exchange}{Unique name of an exchange.}

\item{type}{Enum: \code{"spot"}, \code{"future"}, or "\code{"option"}. Type of markets.}

\item{base}{Base asset of markets.}

\item{quote}{Quote asset of markets.}

\item{asset}{Any asset of markets.}

\item{symbol}{Symbol of derivative markets, full instrument name.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of all implied volatility for option market.
}
\description{
Supported Market Implied Volatility
}
