% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full_v2.R
\name{catalog_full_v2_mempool_feerates}
\alias{catalog_full_v2_mempool_feerates}
\title{Supported Mempool Feerates}
\usage{
catalog_full_v2_mempool_feerates(
  assets = NULL,
  page_size = 2000,
  paging_from = "start"
)
}
\arguments{
\item{assets}{Character vector of assets. By default, all assets are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of all supported assets for the mempool-feerates endpoint,
along with time ranges of available data.
}
\description{
Supported Mempool Feerates
}
