% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full_v2.R
\name{catalog_full_v2_mining_pool_tips_summaries}
\alias{catalog_full_v2_mining_pool_tips_summaries}
\title{Supported Mining Pool Tips Summaries}
\usage{
catalog_full_v2_mining_pool_tips_summaries(
  assets = NULL,
  page_size = 2000,
  paging_from = "start"
)
}
\arguments{
\item{assets}{Character vector of assets. By default, all assets are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of all supported assets for the mining-pool-tips-summary endpoint,
along with time ranges of available data.
}
\description{
Supported Mining Pool Tips Summaries
}
