% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full_v2.R
\name{catalog_full_v2_pair_candles}
\alias{catalog_full_v2_pair_candles}
\title{Supported Pair Candles}
\usage{
catalog_full_v2_pair_candles(
  pairs = NULL,
  page_size = 2000,
  paging_from = "start"
)
}
\arguments{
\item{pairs}{Character vector of asset pairs. By default, all asset pairs are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of all supported asset pair candles,
along with the time ranges of available data per candle duration.
}
\description{
Supported Pair Candles
}
