% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_index_candles}
\alias{catalog_index_candles}
\title{Avilable Index Candles}
\usage{
catalog_index_candles(indexes = NULL)
}
\arguments{
\item{indexes}{Vector of index names. By default, all indexes are returned.}
}
\value{
Tibble of available index candles along with the time ranges of available data per candle duration.
}
\description{
Avilable Index Candles
}
