% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_institution_metrics}
\alias{catalog_institution_metrics}
\title{Available Institution Metrics}
\usage{
catalog_institution_metrics(metrics = NULL, reviewable = NULL)
}
\arguments{
\item{metrics}{Vector of metrics. By default, all metrics are returned.}

\item{reviewable}{Boolean to limit to human-reviewable metrics. By default, all metrics are returned.}
}
\value{
Available institution metrics along with information for them like description, category, and institutions for which a metric is available.
}
\description{
Available Institution Metrics
}
