% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_market_liquidations}
\alias{catalog_market_liquidations}
\title{Get Available Market Liquidations}
\usage{
catalog_market_liquidations(
  markets = NULL,
  exchange = NULL,
  type = NULL,
  base = NULL,
  quote = NULL,
  asset = NULL,
  symbol = NULL,
  format = "json",
  limit = NULL,
  pretty = FALSE
)
}
\arguments{
\item{markets}{Vector of markets. By default all markets are returned.}

\item{exchange}{Unique name of an exchange.}

\item{type}{Type of markets. Available inputs are \code{spot}, \code{future}, or \code{option}.}

\item{base}{Base asset of markets.}

\item{quote}{Quote asset of markets.}

\item{asset}{Any asset of markets.}

\item{symbol}{Symbol of derivative markets; the full instrument name.}

\item{format}{Format of the response. Supported values are \code{json}, \code{json_stream}. Default is \code{json}.}

\item{limit}{Number of response items.}

\item{pretty}{Human-readable formatting of JSON responses.}
}
\value{
Tibble of markets with liquidations support along with the time ranges of available data.
}
\description{
Get Available Market Liquidations
}
