% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_market_openinterest}
\alias{catalog_market_openinterest}
\title{Available Market Open Interest}
\usage{
catalog_market_openinterest(
  markets = NULL,
  exchange = NULL,
  type = NULL,
  base = NULL,
  quote = NULL,
  asset = NULL,
  symbol = NULL
)
}
\arguments{
\item{markets}{Vector of markets. By default all markets are returned.}

\item{exchange}{Unique name of an exchange.}

\item{type}{Type of markets. Available inputs are \code{spot}, \code{future}, or \code{option}.}

\item{base}{Base asset of markets.}

\item{quote}{Quote asset of markets.}

\item{asset}{Any asset of markets.}

\item{symbol}{Symbol of derivative markets; the full instrument name.}
}
\value{
Tibble of markets with open interest support along with the time ranges of available data.
}
\description{
Available Market Open Interest
}
