% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_markets}
\alias{catalog_markets}
\title{Available Markets}
\usage{
catalog_markets(
  markets = NULL,
  exchange = NULL,
  type = NULL,
  base = NULL,
  quote = NULL,
  asset = NULL,
  symbol = NULL,
  include = NULL,
  exclude = NULL
)
}
\arguments{
\item{markets}{Vector of markets. By default all markets are returned.}

\item{exchange}{Unique name of an exchange.}

\item{type}{Type of markets. Available inputs are \code{spot}, \code{future}, or \code{option}.}

\item{base}{Base asset of markets.}

\item{quote}{Quote asset of markets.}

\item{asset}{Any asset of markets.}

\item{symbol}{Symbol of derivative markets; the full instrument name.}

\item{include}{Vector of fields to include in response.
Supported values are \code{trades}, \code{orderbooks}, \code{quotes}, \code{funding_rates}, \code{openinterest}, \code{liquidations}.
Included by default if omitted.}

\item{exclude}{Vector of fields to exclude from response.
Supported values are \code{trades}, \code{orderbooks}, \code{quotes}, \code{funding_rates}, \code{openinterest}, \code{liquidations}.
Included by default if omitted.}
}
\value{
List of available markets along with time ranges of available data.
}
\description{
Available Markets
}
