% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_pair_metrics}
\alias{catalog_pair_metrics}
\title{Get Available Pair Metrics}
\usage{
catalog_pair_metrics(metrics = NULL, as_list = FALSE)
}
\arguments{
\item{metrics}{Vector of metrics. By default all metrics are returned.}

\item{as_list}{Return response as list instead of tibble.}
}
\value{
Available pair metrics along with information for them like descripion, category, and pairs for which a metric is available.
}
\description{
Get Available Pair Metrics
}
