% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_v2.R
\name{catalog_v2_asset_chains}
\alias{catalog_v2_asset_chains}
\title{Available Asset Chains}
\usage{
catalog_v2_asset_chains(assets = NULL, page_size = 2000, paging_from = "start")
}
\arguments{
\item{assets}{Character vector of assets. By default, all assets are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of available assets for the asset-chains endpoint,
along with time ranges of available data.
}
\description{
Available Asset Chains
}
