% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_v2.R
\name{catalog_v2_exchange_asset_metrics}
\alias{catalog_v2_exchange_asset_metrics}
\title{Available Exchange-Asset Metrics}
\usage{
catalog_v2_exchange_asset_metrics(
  exchange_assets = NULL,
  metrics = NULL,
  page_size = 2000,
  paging_from = "start"
)
}
\arguments{
\item{exchange_assets}{Character vector of exchange-assets. By default, all exchange-asset pairs are returned.}

\item{metrics}{Vector of metrics. By default, all metrics are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of available exchange-asset metrics, along with the time ranges of available data.
}
\description{
Available Exchange-Asset Metrics
}
