% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_v2.R
\name{catalog_v2_institution_metrics}
\alias{catalog_v2_institution_metrics}
\title{Available Institutions Metrics}
\usage{
catalog_v2_institution_metrics(
  institutions = NULL,
  metrics = NULL,
  page_size = 2000,
  paging_from = "start"
)
}
\arguments{
\item{institutions}{Character vector of institutions. By default, all institutions are returned.}

\item{metrics}{Vector of metrics. By default, all metrics are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of institutions, along with the time ranges of available data.
}
\description{
Available Institutions Metrics
}
