% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_metrics.R
\name{get_asset_chains}
\alias{get_asset_chains}
\title{Asset chains (experimental)}
\usage{
get_asset_chains(
  assets,
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = 10000,
  paging_from = "end"
)
}
\arguments{
\item{assets}{character vector of assets. Use "*" to get metrics for all supported assets.}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}
}
\value{
The Chains of blocks for specified assets
}
\description{
Asset chains (experimental)
}
