% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_metrics.R
\name{get_asset_metrics}
\alias{get_asset_metrics}
\title{Get Coin Metrics API Data}
\usage{
get_asset_metrics(
  assets,
  metrics,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_height = NULL,
  end_height = NULL,
  start_hash = NULL,
  end_hash = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  min_confirmations = NULL,
  timezone = "UTC",
  page_size = 10000,
  paging_from = "end",
  sort = "asset",
  status = NULL,
  limit_per_asset = NULL,
  null_as_zero = FALSE
)
}
\arguments{
\item{assets}{character vector of assets. Use "*" to get metrics for all supported assets.}

\item{metrics}{Vector of metrics to request}

\item{frequency}{Frequency of metrics. Supported values are \verb{1b} (block by block), \verb{1s}, \verb{1m}, \verb{10m}, \verb{1h}, \verb{1d}, \verb{1d-ny-close}.}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_height}{The beginning block height for the set of data returned. Mutually exclusive with \code{start_time} and \code{start_hash}.}

\item{end_height}{The ending block height for the set of data returned. Mutually exclusive with \code{end_time} and \code{end_hash}.}

\item{start_hash}{The beginning block height for the set of data returned. Mutually exclusive with \code{start_time} and \code{start_height}.}

\item{end_hash}{The ending block height for the set of data returned. Mutually exclusive with \code{end_time} and \code{end_height}.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{min_confirmations}{[0, 1, ... 99] How many blocks behind the block by block metrics (\verb{1b} frequency) are the data based}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}

\item{sort}{Sort tabular data by asset, time, or metric}

\item{status}{String for which metric values you want to see: "all", "flash", "reviewed", "revised"}

\item{limit_per_asset}{Number of entries per asset}

\item{null_as_zero}{Represent nulls as zeros in the response (TRUE or FALSE)}
}
\description{
Parses API response into tibble of metrics data
}
