% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_asset_profiles}
\alias{get_asset_profiles}
\title{Asset profiles data (experimental)}
\usage{
get_asset_profiles(
  assets = NULL,
  full_names = NULL,
  page_size = 1000,
  paging_from = "start"
)
}
\arguments{
\item{assets}{Vector of assets.}

\item{full_names}{Vector of asset full names. Mutually exclusive with \code{assets} parameter.}

\item{page_size}{Number of items per single page of results.}

\item{paging_from}{where the first page starts:  \code{start} or \code{end}.}
}
\value{
Profile data for assets. If \code{supply_cap} is not present then the theoretical maximum supply is infinite for that asset.
}
\description{
Asset profiles data (experimental)
}
