% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{get_block_settlements}
\alias{get_block_settlements}
\title{Block Settlements}
\usage{
get_block_settlements(asset)
}
\arguments{
\item{asset}{Asset name.}
}
\value{
Tibble of block settlements
}
\description{
Block Settlements
}
