% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_all.R
\name{get_catalog_all_assets}
\alias{get_catalog_all_assets}
\title{Supported Assets}
\usage{
get_catalog_all_assets(
  assets = NULL,
  include = NULL,
  exclude = NULL,
  pretty = FALSE
)
}
\arguments{
\item{assets}{Vector of assets. By default all assets are returned.}

\item{include}{Vector of fields to include in response. Supported values are \code{metrics}, \code{markets}, \code{exchanges}.
Included by default if omitted.,}

\item{exclude}{Vector of fields to exclude in response. Supported values are \code{metrics}, \code{markets}, \code{exchanges}.
Included by default if omitted.}

\item{pretty}{Human-readable formatting of JSON responses.}
}
\value{
Tibble of supported assets along with information for them like metrics, markets, exchanges, and time ranges of available data.
Metrics, markets, and exchanges outputted as list-columns.
}
\description{
Supported Assets
}
