% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogs.R
\name{get_catalog_asset_alerts}
\alias{get_catalog_asset_alerts}
\title{Available Asset Alerts}
\usage{
get_catalog_asset_alerts(assets = NULL, alert = NULL)
}
\arguments{
\item{assets}{Vector of asets. By default all assets are returned.}

\item{alerts}{Vector of asset alert names. By default all asset alerts are returned.}
}
\value{
available asset alerts along with their descriptions, thresholds, and constituents.
}
\description{
Available Asset Alerts
}
