% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogs.R
\name{get_catalog_asset_metrics}
\alias{get_catalog_asset_metrics}
\title{Get Available Asset Metrics}
\usage{
get_catalog_asset_metrics(
  metrics = NULL,
  reviewable = TRUE,
  pretty = FALSE,
  as_list = TRUE
)
}
\arguments{
\item{metrics}{Vector of metrics. By default, all metrics are returned.}

\item{reviewable}{Boolean to limit to human-reviewable metrics. By default, all metrics are returned.}

\item{pretty}{Human-readable formatting of JSON responses.}

\item{as_list}{Return list instead of tibble.}
}
\value{
Tibble of available asset metrics along with information like description, category, and assets for which a metric is available.
}
\description{
Get Available Asset Metrics
}
