% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogs.R
\name{get_catalog_asset_pairs}
\alias{get_catalog_asset_pairs}
\title{Available Asset Pairs}
\usage{
get_catalog_asset_pairs(pairs = NULL)
}
\arguments{
\item{pairs}{Vector of asset pairs. By default, all asset pairs are returned.}
}
\value{
List of available asset pairs, along with key information like metrics and time ranges of available data.
}
\description{
Available Asset Pairs
}
