% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{get_catalog_assets}
\alias{get_catalog_assets}
\title{Available Assets}
\usage{
get_catalog_assets(assets = NULL, include = NULL, exclude = NULL)
}
\arguments{
\item{assets}{Vector of assets. By default all assets are returned.}

\item{include}{Vector of fields to include in response. Supported values are \code{metrics}, \code{markets}, \code{exchanges}. Included by default if \code{NULL}.}

\item{exclude}{Vector of fields to exclude from response. Supported values are \code{metrics}, \code{markets}, \code{exchanges}. Included by default if \code{NULL}.}
}
\value{
List of available assets along with information for them (metrics, markets, and exchanges) and time ranges of available data.
}
\description{
Available Assets
}
