% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogs.R
\name{get_catalog_exchange_assets}
\alias{get_catalog_exchange_assets}
\title{Get Available Exchange/Asset Pairs}
\usage{
get_catalog_exchange_assets(
  exchange_assets = NULL,
  pretty = FALSE,
  as_list = TRUE
)
}
\arguments{
\item{exchange_assets}{Vector of exchange-assets. By default, all exchange-asset pairs are returned.}

\item{pretty}{Human-readable formatting of JSON responses.}

\item{as_list}{Return content as list instead of tibble.}
}
\value{
Tibble of Available exchange-asset pairs along with information like metrics and time ranges of available data.
}
\description{
Get Available Exchange/Asset Pairs
}
