% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogs.R
\name{get_catalog_exchanges}
\alias{get_catalog_exchanges}
\title{Get Available Exchanges}
\usage{
get_catalog_exchanges(exchanges = NULL, as_list = FALSE)
}
\arguments{
\item{exchanges}{Vector of exchanges. By default all exchanges are returned.}

\item{as_list}{Return content as list or tibble.}
}
\value{
Available exchanges along with available markets and metrics for them.
}
\description{
Get Available Exchanges
}
