% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{get_catalog_indexes}
\alias{get_catalog_indexes}
\title{Get Available Indexes}
\usage{
get_catalog_indexes(indexes = NULL)
}
\arguments{
\item{indexes}{Vector of index names. By default, all indexes are returned.}
}
\value{
Tibble of available indexes along with time ranges of available data
}
\description{
Get Available Indexes
}
