% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{get_catalog_institution_metrics}
\alias{get_catalog_institution_metrics}
\title{Get Available Institution Metrics}
\usage{
get_catalog_institution_metrics(metrics = NULL, as_list = FALSE)
}
\arguments{
\item{metrics}{Vector of metrics. By default all metrics are returned.}
}
\value{
Available institution metrics along with information for them like description, category, and institutions for which a metric is available.
}
\description{
Get Available Institution Metrics
}
