% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogs.R
\name{get_catalog_markets}
\alias{get_catalog_markets}
\title{Get Available Markets}
\usage{
get_catalog_markets(
  markets = NULL,
  exchange = NULL,
  type = NULL,
  base = NULL,
  quote = NULL,
  asset = NULL,
  symbol = NULL,
  include = NULL,
  exclude = NULL,
  format = "json",
  limit = NULL,
  pretty = FALSE
)
}
\arguments{
\item{markets}{Vector of markets. By default all markets are returned.}

\item{exchange}{Unique name of an exchange.}

\item{type}{Type of markets. Available inputs are \code{spot}, \code{future}, or \code{option}.}

\item{base}{Base asset of markets.}

\item{quote}{Quote asset of markets.}

\item{asset}{Any asset of markets.}

\item{symbol}{Symbol of derivative markets; the full instrument name.}

\item{include}{Vector of fields to include in response.
Supported values are \code{trades}, \code{orderbooks}, \code{quotes}, \code{funding_rates}, \code{openinterest}, \code{liquidations}.
Included by default if omitted.}

\item{exclude}{Vector of fields to exclude from response.
Supported values are \code{trades}, \code{orderbooks}, \code{quotes}, \code{funding_rates}, \code{openinterest}, \code{liquidations}.
Included by default if omitted.}

\item{format}{Format of the response. Supported values are \code{json}, \code{json_stream}. Default is \code{json}.}

\item{limit}{Number of response items.}

\item{pretty}{Human-readable formatting of JSON responses.}
}
\value{
List of available markets along with time ranges of available data.
}
\description{
Get Available Markets
}
