% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{get_catalog_pair_candles}
\alias{get_catalog_pair_candles}
\title{Get Available Asset Pair Candles}
\usage{
get_catalog_pair_candles(pairs = NULL, pretty = FALSE, as_list = FALSE)
}
\arguments{
\item{pairs}{Vector of asset pairs. By default, all asset pairs are returned.}

\item{pretty}{Human-readable formatting of JSON responses.}
}
\value{
Tibble of available asset pair candles along with the time ranges of available data per candle duration.
}
\description{
Get Available Asset Pair Candles
}
