% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalogs.R
\name{get_catalog_pair_metrics}
\alias{get_catalog_pair_metrics}
\title{Get Available Pair Metrics}
\usage{
get_catalog_pair_metrics(metrics = NULL, as_list = FALSE)
}
\arguments{
\item{metrics}{Vector of metrics. By default all metrics are returned.}
}
\value{
Available pair metrics along with information for them like descripion, category, and pairs for which a metric is available.
}
\description{
Get Available Pair Metrics
}
