% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{get_catalog_pairs}
\alias{get_catalog_pairs}
\title{Available Asset Pairs}
\usage{
get_catalog_pairs(pairs = NULL, as_list = TRUE)
}
\arguments{
\item{pairs}{Vector of asset pairs. By default, all asset pairs are returned.}

\item{as_list}{Return list instead of tibble.}
}
\value{
List of available asset pairs, along with key information like metrics and time ranges of available data.
}
\description{
Available Asset Pairs
}
