% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_util.R
\name{get_coinmetrics_api_data}
\alias{get_coinmetrics_api_data}
\title{Get API Data}
\usage{
get_coinmetrics_api_data(api_response, endpoint, paging_from, as_list = NULL)
}
\arguments{
\item{api_response}{API response as received from \code{httr::GET}}

\item{endpoint}{Corresponding API endpoint (i.e. \code{market-candles} or \code{index-levels})}

\item{paging_from}{Enum: (\code{start}, \code{end}) Corresponding pagination in original request.}

\item{as_list}{Deprecated, included only for compatibility, to be removed later.}
}
\value{
Data table
}
\description{
Get API Data
}
