% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defi.R
\name{get_defi_balance_sheets}
\alias{get_defi_balance_sheets}
\title{DeFi Balance Sheets (experimental)
Data is updated on an end-of-day frequency}
\usage{
get_defi_balance_sheets(
  defi_protocols,
  start_time = NULL,
  end_time = NULL,
  start_height = NULL,
  end_height = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  pretty = FALSE
)
}
\arguments{
\item{defi_protocols}{Vector of DeFi protocols like \code{aave_v2_eth} or protocol patterns \verb{aave_v2_*} or \verb{aave_*_eth} or \verb{*_eth}}

\item{start_time}{Start of the time interval. If omitted, response starts from the earliest time available.}

\item{end_time}{End of the time interval. If omitted, response includes data up to the latest time available.}

\item{start_height}{The beginning block height for the set of data returned.}

\item{end_height}{The ending block height for the set of data returned.}

\item{start_inclusive}{Inclusive or exclusive corresponding \verb{start_*} parameters.}

\item{end_inclusive}{Inclusive or exclusive corresponding \verb{end_*} parameters.}

\item{timezone}{Timezone name for \code{start_time} and \code{end_time} timestamps. Note, the output times are always in \code{UTC}.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where the fist page starts, at the beginning or end of the interval.}

\item{pretty}{Human-readable formatting of JSON responses.}
}
\value{
List of DeFi balance sheet records for specified DeFi protocols
}
\description{
DeFi Balance Sheets (experimental)
Data is updated on an end-of-day frequency
}
