% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_metrics.R
\name{get_exchange_asset_metrics}
\alias{get_exchange_asset_metrics}
\title{Exchange-asset Metrics}
\usage{
get_exchange_asset_metrics(
  exchange_assets,
  metrics,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  sort = "exchange_asset",
  limit_per_exchange_asset = NULL,
  pretty = FALSE,
  format = "json",
  as_list = FALSE
)
}
\arguments{
\item{exchange_assets}{Vector of exchange-asset pairs, e.g. binance-btc}

\item{metrics}{Vector of metrics to request. Information on all available metrics can be found on page https://docs.coinmetrics.io/info/pair_metrics.}

\item{frequency}{Frequency of exchange-asset metrics. Supported values are \verb{5m}, \verb{1h}, \verb{1d}.}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}

\item{sort}{How results will be sorted. Metrics are sorted by \verb{(exchange_asset, time)} by default. Posible values are \emph{exchange_asset} or \emph{time}.}

\item{limit_per_exchange_asset}{Number of entries per exchange_asset.}

\item{pretty}{Human-readable JSON formatting. Default is false.}

\item{format}{Format of the response, \code{json} or \code{csv}. Default is \code{json}.}

\item{as_list}{Should API return list instead of tabular data}
}
\value{
Tibble or list of metrics for specified exchange-asset.
}
\description{
Exchange-asset Metrics
}
