% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_metrics.R
\name{get_exchange_metrics}
\alias{get_exchange_metrics}
\title{Exchange Metrics}
\usage{
get_exchange_metrics(
  exchanges,
  metrics,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  sort = "exchange",
  limit_per_exchange = NULL,
  pretty = FALSE,
  format = "json",
  as_list = FALSE
)
}
\arguments{
\item{exchanges}{Vector of exchange names.}

\item{metrics}{Vector of metrics to request. Information on all available metrics can be found on page https://docs.coinmetrics.io/info/pair_metrics.}

\item{frequency}{Frequency of the pair metrics. Supported values are \verb{1h} and \verb{1d}}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}

\item{limit_per_exchange}{Number of entries per institution.}

\item{pretty}{Human-readable JSON formatting. Default is false.}

\item{as_list}{Should API return list instead of tabular data}
}
\value{
Tibble or list of metrics for specified exchanges. Results are ordered by tuple (exchange, time).
}
\description{
Exchange Metrics
}
