% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{get_full_block}
\alias{get_full_block}
\title{Full block}
\usage{
get_full_block(asset, block_hash)
}
\arguments{
\item{asset}{Asset name}

\item{block_hash}{Block hash}
}
\value{
Full blockchain blok with all transactions and balance updates.
The result is a list of threee tibbles. The first element, \code{metadata}, contains summary information for the block.
The second element, \code{block_transactions} list the transactions for the block.
The third element, \code{balance_updates} lists the balance updates the block.
}
\description{
Full block
}
