% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{get_full_transaction}
\alias{get_full_transaction}
\title{Get Full Transaction}
\usage{
get_full_transaction(asset, txid)
}
\arguments{
\item{asset}{Asset name.}

\item{txid}{Transaction identifier.}
}
\value{
Full blockchain transaction with all balance updates.
The result is a list of two tibbles. The first element, \code{metadata}, contains summary information for the transaction.
The second element, \code{balance_updates} lists the balance updates for the transaction.
}
\description{
Get Full Transaction
}
