% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{get_full_transaction_for_block}
\alias{get_full_transaction_for_block}
\title{Full Transaction for Block}
\usage{
get_full_transaction_for_block(asset, block_hash, txid)
}
\arguments{
\item{asset}{Asset name}

\item{block_hash}{Block hash}

\item{txid}{Transaction identifier (txid)}
}
\value{
Full blockchain transaction with all balance updates for a specific block.
The result is a list of two tibbles. The first element, \code{metadata}, contains summary information for the transaction.
The second element, \code{balance_updates} lists the balance updates for the transaction.
}
\description{
Full Transaction for Block
}
