% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indexes.R
\name{get_index_constituents}
\alias{get_index_constituents}
\title{Get Index Constituents}
\usage{
get_index_constituents(
  indexes,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = 10000,
  paging_from = "end"
)
}
\arguments{
\item{indexes}{vector of index names, e.g. 'CMBI10'}

\item{frequency}{Frequency of the index. Supported values are 1s, 15s, 1h, 1d-ny-close, 1d-sg-close, 1d, 1d-HH:00}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}
}
\value{
Tibble of constituents and weights for specified indexes.
}
\description{
Get Index Constituents
}
