% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_institution_metrics.R
\name{get_institution_metrics}
\alias{get_institution_metrics}
\title{Get Institution Metrics}
\usage{
get_institution_metrics(
  institutions,
  metrics,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  sort = "institution",
  limit_per_institution = NULL,
  pretty = FALSE,
  format = "json",
  as_list = FALSE
)
}
\arguments{
\item{institutions}{Vector of institutions.}

\item{metrics}{Vector of metrics to request time series data.}

\item{frequency}{Frequency of the institution metrics. Supported values are \verb{1d}.}

\item{start_time}{Start of the time interval.}

\item{end_time}{End of the time interval.}

\item{start_inclusive}{Inclusive or exclusive corresponding \verb{start_*} parameters.}

\item{end_inclusive}{Inclusive or exclusive corresponding \verb{end_*} parameters.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Example: America/New_York.}

\item{page_size}{Number of items per single page of results.}

\item{paging_from}{First page start at the start or end of the interval.}

\item{sort}{How results will be sorted; sorted by \verb{(institution, time)} by default.}

\item{limit_per_institution}{Number of entries per institution.}

\item{pretty}{Human-readable formatting of JSON responses.}

\item{format}{Either \code{"json"} or "\code{"csv"} response format.}

\item{as_list}{Return data as list.}
}
\value{
Tibble or list of metrics for specified institutions.
}
\description{
Get Institution Metrics
}
