% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain_v2.R
\name{get_list_of_accounts_v2}
\alias{get_list_of_accounts_v2}
\title{Get List of Accounts}
\usage{
get_list_of_accounts_v2(
  asset,
  accounts = NULL,
  start_time = NULL,
  end_time = NULL,
  start_height = NULL,
  end_height = NULL,
  start_chain_sequence_number = NULL,
  end_chain_sequence_number = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end"
)
}
\arguments{
\item{asset}{Asset name.}

\item{accounts}{Optional vector of accounts to filter a response.
This \emph{must} contain a single element for Community API users.}

\item{start_time}{Start of the time interval. Mutually exclusive with \code{start_height}.
If omitted, response will include time series from the \strong{earliest} time available.
This parameter is disabled for Community API users.}

\item{end_time}{End of the time interval. Mutually exclusive with \code{end_height}.
If omitted, response will include time series up to the \strong{latest} time available.
This parameter is disabled for Community API users.}

\item{start_height}{Beginning block height for the set of data returned, inclusive by default.
Mutually exclusive with \code{start_time}. This parameter is disabled for Community API users.}

\item{end_height}{Ending block height for the set of data returned, inclusive by default.
Mutually exclusive with \code{end_time}. This parameter is disabled for Community API users.}

\item{start_chain_sequence_number}{Start of the \code{chain_sequence_number} interval.  This parameter is disabled for Community API users.}

\item{end_chain_sequence_number}{End of the \code{chain_sequence_number} interval. This parameter is disabled for Community API users.}

\item{start_inclusive}{Inclusive or exclusive corresponding \verb{start_*} parameters.
This parameter is disabled for Community API users.}

\item{end_inclusive}{Inclusive or exclusive corresponding \verb{end_*} parameters.
This parameter is disabled for Community API users.}

\item{timezone}{Timezone name (e.g. \code{America/New_York}) for \code{start_time} and \code{end_time} timestamps. Default is \code{UTC}.
This parameter does not modify the output times, which are always UTC.}

\item{page_size}{Number of items per single page of results. Integer between 1 and 10000.
This parameter is disabled for Community API users.}

\item{paging_from}{First page starts at the beginning or end of the interval. Supported values are \code{start} and \code{end}.}
}
\value{
Tibble of blockchain account with their balances.
Results are ordered by tuple \verb{(creation_chain_sequence_number, account)}.
}
\description{
Get List of Accounts
}
