% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain_v2.R
\name{get_list_of_balance_updates_v2}
\alias{get_list_of_balance_updates_v2}
\title{Get List of Balance Updates}
\usage{
get_list_of_balance_updates_v2(
  asset,
  accounts = NULL,
  limit_per_account = NULL,
  sub_accounts = NULL,
  txids = NULL,
  block_hashes = NULL,
  start_time = NULL,
  end_time = NULL,
  start_height = NULL,
  end_height = NULL,
  start_chain_sequence_number = NULL,
  end_chain_sequence_number = NULL,
  include_sub_accounts = NULL,
  chain = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end"
)
}
\arguments{
\item{asset}{Asset name.}

\item{accounts}{Optional vector of accounts to filter a response.
This \emph{must} contain a single element for Community API users.}

\item{limit_per_account}{Number of entries per account the result should contain.  Applicable when multiiple accounts are requested.
For Community API users, this parameter is disabled and the limit is fixed to 100.}

\item{sub_accounts}{Optional vector of sub-accounts to filter a response.
This parameter is disabled for Community API users.}

\item{txids}{Optional vector of transaction identifiers (txid) to filter a response.
For Community API users, this vector must contain a single element.}

\item{block_hashes}{Optional vector of block hashes to filter a response.
This \emph{must} contain a single element for Community API users.}

\item{start_time}{Start of the time interval. Mutually exclusive with \code{start_height}.
If omitted, response will include time series from the \strong{earliest} time available.
This parameter is disabled for Community API users.}

\item{end_time}{End of the time interval. Mutually exclusive with \code{end_height}.
If omitted, response will include time series up to the \strong{latest} time available.
This parameter is disabled for Community API users.}

\item{start_height}{Beginning block height for the set of data returned, inclusive by default.
Mutually exclusive with \code{start_time}. This parameter is disabled for Community API users.}

\item{end_height}{Ending block height for the set of data returned, inclusive by default.
Mutually exclusive with \code{end_time}. This parameter is disabled for Community API users.}

\item{start_chain_sequence_number}{Start of the \code{chain_sequence_number} interval.  This parameter is disabled for Community API users.}

\item{end_chain_sequence_number}{End of the \code{chain_sequence_number} interval. This parameter is disabled for Community API users.}

\item{include_sub_accounts}{Boolean indicating if the response should contain sub-accounts.
This parameter is disabled for Community API users.}

\item{chain}{Chain type. Default is \code{"main"}. Supported values are \code{"main"} and \code{"all"} (includes both main and stale).
This parameter is disabled for Community API users.}

\item{start_inclusive}{Inclusive or exclusive corresponding \verb{start_*} parameters.
This parameter is disabled for Community API users.}

\item{end_inclusive}{Inclusive or exclusive corresponding \verb{end_*} parameters.
This parameter is disabled for Community API users.}

\item{timezone}{Timezone name (e.g. \code{America/New_York}) for \code{start_time} and \code{end_time} timestamps. Default is \code{UTC}.
This parameter does not modify the output times, which are always UTC.}

\item{page_size}{Number of items per single page of results. Integer between 1 and 10000.
This parameter is disabled for Community API users.}

\item{paging_from}{First page starts at the beginning or end of the interval. Supported values are \code{start} and \code{end}.}
}
\value{
List of blockchain accounts balance updates.
}
\description{
Get List of Balance Updates
}
