% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{get_list_of_transactions}
\alias{get_list_of_transactions}
\title{Get List of Transactions}
\usage{
get_list_of_transactions(
  asset,
  transaction_hashes = NULL,
  block_hashes = NULL,
  start_time = NULL,
  end_time = NULL,
  start_height = NULL,
  end_height = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end"
)
}
\arguments{
\item{asset}{Asset name.}

\item{transaction_hashes}{Optional vector of transaction hashes to filter a response.}

\item{block_hashes}{Optional vector of block hashes to filter a response.}

\item{start_time}{Start of the time interval. This field refers to the \code{time} field in the response.}

\item{end_time}{End of the time interval. This field refers to the \code{time} field in the response.}

\item{start_height}{(int >= 0) Beginning block height for the set of data returned.  Inclusive by default. Mutually exclusive with \code{start_time}.}

\item{end_height}{(int >= 0) Ending block height for the set of data returned. Inclusive by default. Mutually exclusive with \code{end_time}.}

\item{start_inclusive}{Inclusive or exclusive corresponding \verb{start_*} parameters.}

\item{end_inclusive}{Inclusive or exclusive corresponding \verb{end_*} parameters.}

\item{timezone}{Timezone name for \code{start_time} and \code{end_time} timestamps, \code{"UTC"} by default. Output times are always \code{UTC}.}

\item{page_size}{Number of items per single page of results.}

\item{paging_from}{First page at start or end of interval. Enum: \code{"start"} or \code{"end"}.}
}
\value{
Tibble of blockchain transactions metadata.
Results are ordered by tuple \verb{(min_chain_sequence_number, transaction_hash)}.
}
\description{
Get List of Transactions
}
