% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candles.R
\name{get_market_candles}
\alias{get_market_candles}
\title{Get Market Candles}
\usage{
get_market_candles(
  markets,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  limit_per_market = NULL,
  pretty = FALSE,
  as_list = FALSE
)
}
\arguments{
\item{markets}{Vector of markets.}

\item{frequency}{Candle duration. Supported values are \verb{1m}, \verb{5m}, \verb{10m}, \verb{15m}, \verb{30m}, \verb{1h}, \verb{4h}, \verb{1d}.}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}

\item{limit_per_market}{Number of entries per market.}

\item{pretty}{Human-readable JSON formatting. Default is false.}

\item{as_list}{Should API return list instead of tabular data}
}
\value{
Tibble or list of candles for specified markets
}
\description{
Get Market Candles
}
