% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_metrics.R
\name{get_market_metrics}
\alias{get_market_metrics}
\title{Get Market Metrics}
\usage{
get_market_metrics(
  markets,
  metrics,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  sort = "market",
  limit_per_market = NULL,
  pretty = FALSE,
  format = "json",
  as_list = FALSE
)
}
\arguments{
\item{markets}{Vector of markets.}

\item{metrics}{Vector of metrics to request time series data}

\item{frequency}{Frequency of the market metrics. Supported values are \verb{1m}, \verb{5m}, \verb{1h}, \verb{1d}.}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}

\item{sort}{How results will be sorted. Metrics are sorted by \verb{(pair, time)} by default.
If you want to sort \verb{1d} metrics by \verb{(time, pair)}, choose \code{sort = "time"}.}

\item{limit_per_market}{Number of entries per market.}

\item{pretty}{Human-readable JSON formatting. Default is false.}

\item{format}{Format of the response, \code{json} or \code{csv}. Default is \code{json}.}

\item{as_list}{Should API return list instead of tabular data}
}
\value{
Tibble or list of metrics for specified markets.
}
\description{
Get Market Metrics
}
