% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mining.R
\name{get_mining_pool_tips}
\alias{get_mining_pool_tips}
\title{Mining pool tips summary}
\usage{
get_mining_pool_tips(
  assets,
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  pretty = TRUE,
  as_list = TRUE
)
}
\arguments{
\item{assets}{character vector of assets. Use "*" to get metrics for all supported assets.}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}

\item{pretty}{Human-readable JSON formatting. Default is false.}

\item{as_list}{Return the response as (nested) list or tibble.}
}
\value{
Last tips we saw in all active mining pools during the past hour. See \href{https://docs.coinmetrics.io/api/v4#operation/getTimeseriesMiningPoolTipsSummary}{API Documentation} for more information.
}
\description{
Mining pool tips summary
}
