% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pair_metrics.R
\name{get_pair_metrics}
\alias{get_pair_metrics}
\title{Pair Metrics}
\usage{
get_pair_metrics(
  pairs,
  metrics,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  sort = "pair",
  limit_per_pair = NULL
)
}
\arguments{
\item{pairs}{Vector of asset pairs, e.g. btc-usd.}

\item{metrics}{Vector of metrics to request, e.g. volume_trusted_spot_usd_1d.}

\item{frequency}{Frequency of the pair metrics. Supported values are \verb{1h} and \verb{1d}.}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}

\item{sort}{How results will be sorted. Metrics are sorted by \verb{(pair, time)} by default.
If you want to sort \verb{1d} metrics by \verb{(time, pair)}, choose \code{sort = "time"}.}

\item{limit_per_pair}{Number of entries per pair the result should contain.}
}
\value{
tibble or list of time series asset pairs
}
\description{
Pair Metrics
}
