% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Taxonomy Endpoints}
\usage{
get_taxonomy(
  assets = NULL,
  class_ids = NULL,
  sector_ids = NULL,
  subsector_ids = NULL,
  version = "latest",
  classification_start_time = NULL,
  classification_end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  page_size = 500,
  paging_from = "start",
  pretty = FALSE
)
}
\arguments{
\item{assets}{vector of assets}

\item{class_ids}{vector of class identifiers}

\item{sector_ids}{vector of sector identifiers}

\item{subsector_ids}{vector of subsector identifiers}

\item{version}{Taxonomy version. Defaults to \code{latest} when no \verb{*_time} parameters specified. Specify asterisk \code{*} to get all versions.}

\item{classification_start_time}{start time of asset classification}

\item{classification_end_time}{end time of asset classification}

\item{start_inclusive}{inclusive or exclusive corresponding \verb{start_*} parameters}

\item{end_inclusive}{inclusive or exclusive corresponding \verb{end_*} parameters}

\item{page_size}{number of items per single page of results}

\item{paging_from}{where the first page starts:  \code{start} or \code{end}}

\item{pretty}{Human-readable formatting of JSON responses}
}
\value{
Taxonomy for assets, ordered by tuple \verb{(asset, classification_start_time)}
}
\description{
Taxonomy Endpoints
}
