% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_taxonomy_metadata}
\alias{get_taxonomy_metadata}
\title{Taxonomy Metadata}
\usage{
get_taxonomy_metadata(
  version = "latest",
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  page_size = NULL,
  paging_from = "start"
)
}
\arguments{
\item{version}{Taxonomy version. Defaults to \code{latest} when no \verb{*_time} parameters specified. Specify asterisk \code{*} to get all versions.}

\item{start_time}{Start time of taxonomy version.}

\item{end_time}{End time of taxonomy version.}

\item{start_inclusive}{inclusive or exclusive corresponding \verb{start_*} parameters}

\item{end_inclusive}{inclusive or exclusive corresponding \verb{end_*} parameters}

\item{page_size}{number of items per single page of results}

\item{paging_from}{where the first page starts:  \code{start} or \code{end}}
}
\value{
Taxonomy metadata for assets
}
\description{
Taxonomy Metadata
}
