% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data.R
\name{reference_data_exchange_metrics}
\alias{reference_data_exchange_metrics}
\title{Exchange Metrics}
\usage{
reference_data_exchange_metrics(
  metrics = NULL,
  page_size = NULL,
  paging_from = "start"
)
}
\arguments{
\item{metrics}{Vector of metrics. By default all metrics are returned.}

\item{page_size}{Number of items per single page of results, default is 100.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of exchange metrics metadata.
}
\description{
Exchange Metrics
}
