% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data.R
\name{reference_data_indexes}
\alias{reference_data_indexes}
\title{Indexes}
\usage{
reference_data_indexes(indexes = NULL, page_size = NULL, paging_from = "start")
}
\arguments{
\item{indexes}{Character vector of indexes. By default, all indexes are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of indexes metadata.
}
\description{
Indexes
}
