% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data.R
\name{reference_data_pairs}
\alias{reference_data_pairs}
\title{Pairs}
\usage{
reference_data_pairs(pairs = NULL, page_size = NULL, paging_from = "start")
}
\arguments{
\item{pairs}{Character vector of asset pairs. By default, all asset pairs are returned.}

\item{page_size}{Number of items per single page of results, default is \code{100}.}

\item{paging_from}{Enum: \code{start} or \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of pairs metadata.
}
\description{
Pairs
}
