% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_master.R
\name{security_master_assets}
\alias{security_master_assets}
\title{Security Master Assets}
\usage{
security_master_assets(
  assets = NULL,
  codes = NULL,
  page_size = NULL,
  paging_from = "start"
)
}
\arguments{
\item{assets}{Character vector of assets to query. Mutually exclusive with \code{codes}.}

\item{codes}{Character vector of ten-digit alphanumeric identifying codes. Mutually exclusive with \code{assets}.}

\item{page_size}{Integer number of items per single page of results. Default is 100.}

\item{paging_from}{Enum: \code{start}, \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of all assets and their metadata in security master.
}
\description{
Security Master Assets
}
