% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_master.R
\name{security_master_markets}
\alias{security_master_markets}
\title{Security Master Markets}
\usage{
security_master_markets(
  type = NULL,
  markets = NULL,
  symbol = NULL,
  exchange = NULL,
  base = NULL,
  quote = NULL,
  page_size = NULL,
  paging_from = "start"
)
}
\arguments{
\item{type}{Enum: \code{spot}, \code{future}, \code{option}.Type of markets.
If \code{type} is NULL, then ALL markets are selected, and the API response could become unreasonably large.}

\item{markets}{Character vector of markets.}

\item{symbol}{Symbol of derivative markets, full instrument name.}

\item{exchange}{Unique name of an exchange.}

\item{base}{Base asset of markets.}

\item{quote}{Quote asset of markets.}

\item{page_size}{Integer number of items per single page of results. Default is 100.}

\item{paging_from}{Enum: \code{start}, \code{end}. Where does the first page start, at the start of the interval or at the end.}
}
\value{
Data table of metadata on all the markets offered (spot, options, futures), sorted alphabetically by market.
}
\description{
Security Master Markets
}
