% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{set.app.token}
\alias{set.app.token}
\title{Set an Adjust app.token once using this function and you can issue multiple API requests saving yourself having to
pass the token every time.}
\usage{
set.app.token(app.token)
}
\description{
Set an Adjust app.token once using this function and you can issue multiple API requests saving yourself having to
pass the token every time.
}
\seealso{
\code{\link{set.user.token}}, \code{\link{user.token}}, \code{\link{adjust.setup}}, \code{\link{app.token}}
}

