% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{set.user.token}
\alias{set.user.token}
\title{Set an Adjust user.token, required for authorization.}
\usage{
set.user.token(user.token)
}
\description{
Set an Adjust user.token, required for authorization.
}
\seealso{
\code{\link{adjust.setup}}, \code{\link{user.token}}, \code{\link{set.app.token}}, \code{\link{app.token}}
}

