% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{adjust.setup}
\alias{adjust.setup}
\title{Convenience function for initiating a session with a user token and app token, generally required for the start of an Adjust API
session.}
\usage{
adjust.setup(user.token, app.token = NULL)
}
\description{
Convenience function for initiating a session with a user token and app token, generally required for the start of an Adjust API
session.
}
\seealso{
\code{\link{set.user.token}}, \code{\link{user.token}}, \code{\link{set.app.token}}, \code{\link{app.token}}
}

