% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_exchanges}
\alias{catalog_full_exchanges}
\title{Supported Exchanges}
\usage{
catalog_full_exchanges(exchanges = NULL)
}
\arguments{
\item{exchanges}{Vector of exchanges. By default, all exchanges are returned.}
}
\value{
Tibble of all supported exchanges along with available markets for them.
}
\description{
Supported Exchanges
}
