% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_institutions}
\alias{catalog_full_institutions}
\title{Supported Institutions}
\usage{
catalog_full_institutions(institutions = NULL)
}
\arguments{
\item{institutions}{Vector of institutions. By default all institutions are returned.}
}
\value{
Tibble of supported institutions along with information for them like metrics and time ranges of available data.
}
\description{
Supported Institutions
}
