% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_full.R
\name{catalog_full_markets}
\alias{catalog_full_markets}
\title{Supported Markets}
\usage{
catalog_full_markets(
  markets = NULL,
  exchange = NULL,
  type = NULL,
  base = NULL,
  quote = NULL,
  asset = NULL,
  symbol = NULL,
  include = NULL,
  exclude = NULL,
  format = "json",
  limit = NULL,
  pretty = FALSE
)
}
\arguments{
\item{markets}{Vector of markets. By default all markets are returned.}

\item{exchange}{Unique name of an exchange.}

\item{type}{Type of markets. Available inputs are \code{spot}, \code{future}, or \code{option}.}

\item{base}{Base asset of markets.}

\item{quote}{Quote asset of markets.}

\item{asset}{Any asset of markets.}

\item{symbol}{Symbol of derivative markets; the full instrument name.}

\item{include}{Vector of fields to include in response.
Supported values are \code{trades}, \code{orderbooks}, \code{quotes}, \code{funding_rates}, \code{openinterest}, \code{liquidations}.
Included by default if omitted.}

\item{exclude}{Vector of fields to exclude from response.
Supported values are \code{trades}, \code{orderbooks}, \code{quotes}, \code{funding_rates}, \code{openinterest}, \code{liquidations}.
Included by default if omitted.}
}
\value{
List of supported markets along with time ranges of available data.
}
\description{
Supported Markets
}
