% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_pair_candles}
\alias{catalog_pair_candles}
\title{Available Asset Pair Candles}
\usage{
catalog_pair_candles(pairs = NULL)
}
\arguments{
\item{pairs}{Vector of asset pairs. By default, all asset pairs are returned.}
}
\value{
Tibble of available asset pair candles along with the time ranges of available data per candle duration.
}
\description{
Available Asset Pair Candles
}
