% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_pair_metrics}
\alias{catalog_pair_metrics}
\title{Available Pair Metrics}
\usage{
catalog_pair_metrics(metrics = NULL, reviewable = NULL)
}
\arguments{
\item{metrics}{Vector of metrics. By default, all metrics are returned.}

\item{reviewable}{Boolean to limit to human-reviewable metrics. By default, all metrics are returned.}
}
\value{
Available pair metrics along with information for them like descripion, category, and pairs for which a metric is available.
}
\description{
Available Pair Metrics
}
