% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmetrics-package.R
\docType{package}
\name{coinmetrics-package}
\alias{coinmetrics}
\alias{coinmetrics-package}
\title{coinmetrics: An R Interface to Coin Metrics API}
\description{
This package provides an R client to Coin Metrics API.
}
\author{
\strong{Maintainer}: Dorian Kandi \email{dorian.kandi@coinmetrics.io} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
