% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_metrics.R
\name{get_asset_metrics}
\alias{get_asset_metrics}
\title{Get Coin Metrics API Data}
\usage{
get_asset_metrics(
  assets,
  metrics,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_height = NULL,
  end_height = NULL,
  start_hash = NULL,
  end_hash = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  min_confirmations = NULL,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  sort = "asset",
  status = NULL,
  limit_per_asset = NULL,
  pretty = FALSE,
  format = "json",
  next_page_token = NULL,
  pagination = TRUE,
  null_as_zero = NULL
)
}
\arguments{
\item{assets}{character vector of assets. Use "*" to get metrics for all supported assets.}

\item{metrics}{}

\item{frequency}{}

\item{start_time}{}

\item{end_time}{}

\item{start_height}{}

\item{end_height}{}

\item{start_hash}{}

\item{end_hash}{}

\item{start_inclusive}{}

\item{end_inclusive}{}

\item{min_confirmations}{}

\item{timezone}{}

\item{page_size}{}

\item{paging_from}{}

\item{sort}{}

\item{status}{}

\item{limit_per_asset}{}

\item{pretty}{}

\item{format}{}

\item{next_page_token}{}

\item{pagination}{}

\item{null_as_zero}{}
}
\description{
Parses API response into tibble of metrics data
}
