% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_metrics.R
\name{get_asset_metrics}
\alias{get_asset_metrics}
\title{Get Coin Metrics API Data}
\usage{
get_asset_metrics(
  assets,
  metrics,
  frequency = "1d",
  start_time = NULL,
  end_time = NULL,
  start_height = NULL,
  end_height = NULL,
  start_hash = NULL,
  end_hash = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  min_confirmations = NULL,
  timezone = "UTC",
  page_size = 100,
  paging_from = "end",
  sort = "asset",
  status = NULL,
  limit_per_asset = NULL,
  pretty = FALSE,
  format = "json",
  null_as_zero = NULL,
  as_list = FALSE
)
}
\arguments{
\item{assets}{character vector of assets. Use "*" to get metrics for all supported assets.}

\item{metrics}{}

\item{frequency}{}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_height}{}

\item{end_height}{}

\item{start_hash}{}

\item{end_hash}{}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{min_confirmations}{}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}

\item{sort}{Sort tabular data by asset, time, or metric}

\item{status}{}

\item{limit_per_asset}{}

\item{pretty}{Human-readable JSON formatting. Default is false.}

\item{format}{JSON or CSV}

\item{null_as_zero}{}

\item{as_list}{Should API return list instead of tabular data}
}
\description{
Parses API response into tibble of metrics data
}
