% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_trades.R
\name{get_market_trades}
\alias{get_market_trades}
\title{Get Market Trades}
\usage{
get_market_trades(
  markets,
  start_time = NULL,
  end_time = NULL,
  start_inclusive = TRUE,
  end_inclusive = TRUE,
  timezone = "UTC",
  page_size = NULL,
  paging_from = "end",
  limit_per_market = NULL,
  pretty = FALSE,
  format = "json",
  min_confirmations = NULL,
  as_list = FALSE
)
}
\arguments{
\item{markets}{Vector of markets.}

\item{start_time}{Start time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{end_time}{End time of the timeseries. Multiple formats of ISO 8601 are supported.}

\item{start_inclusive}{Flag to define if start timestamp must be included in the timeseries if present. True by default.}

\item{end_inclusive}{Flag to define if end timestamp must be included in the timeseries if present. True by default.}

\item{timezone}{Timezone for \code{start_time} and \code{end_time} timestamps. Default is UTC. Example: America/New_York.}

\item{page_size}{Number of items per single page of results. Default is 100.}

\item{paging_from}{Where first page starts: start or end of interval.}

\item{limit_per_market}{Number of entries per market.}

\item{pretty}{Human-readable JSON formatting. Default is false.}

\item{format}{Format of the response, \code{json} or \code{csv}. Default is \code{json}.}

\item{min_confirmations}{Integer in \verb{[0, 99]} with 2 as default. Specifies how many blocks behind the chain tip trades are based on.
Currently only available for DeFi markets.}

\item{as_list}{Should API return list instead of tabular data}
}
\value{
Tibble or list of selected market trades.
}
\description{
Get Market Trades
}
