% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{user.token}
\alias{user.token}
\title{Get the currently set authorization user.token.}
\usage{
user.token()
}
\description{
Get the currently set authorization user.token.
}
\seealso{
\code{\link{set.user.token}}, \code{\link{adjust.setup}}, \code{\link{set.app.token}}, \code{\link{app.token}}
}

